+++ rasa_model_report/helpers/constants.py
@@ -1,5 +1,5 @@
 DISABLE_NLU = False
-GRADE_PRECISION = 2
+SCORE_PRECISION = 2
 NO_IMAGES = False
 OUTPUT_PATH = "./"
 PROJECT_NAME = "My project"
@@ -8,3 +8,4 @@ RASA_API_URL = "http://localhost:5005"
 RASA_PATH = "./"
 RASA_VERSION = None
 EXCLUDE = []
+VERSION = "1.4.2b11"